feat: protecting delete method for contacts

+++ controllers/contactController.js
@@ -77,6 +77,10 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts");
+    }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json({message: `Delete contact for ${req.params.id}`});
 });
